requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
PyYAML==6.0.1

# Database
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Caching
redis==5.0.1
aioredis==2.0.1
celery>=5.3.6
flower>=2.0.1

# Machine Learning & Analytics
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0
prophet==1.1.4

# Monitoring & Metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# HTTP Client
httpx==0.25.1
aiohttp==3.9.1

# Security
python-dotenv==1.0.0
cryptography==41.0.5
itsdangerous==2.1.2
pyjwt==2.8.0

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock>=3.12.0
httpx==0.25.1

# Development Tools
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1
pre-commit==3.5.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.14
mkdocstrings>=0.23.0

# Utilities
tenacity==8.2.3
python-dateutil==2.8.2
pytz==2023.3.post1
tqdm==4.66.1
loguru==0.7.2

# Monitoring and metrics
sentry-sdk[fastapi]==1.40.0
psutil==5.9.8
cachetools==5.3.2
asyncpg==0.29.0
aioredis==2.0.1
limits==3.9.0

# HTTP and async
websockets==12.0
aiofiles==23.2.1
aiosmtplib==3.0.1

# ML and data processing
scipy>=1.12.0
statsmodels>=0.14.1
pyarrow>=14.0.1
joblib>=1.3.2
tqdm>=4.66.1

# Deep learning
keras==2.15.0
tensorflow-hub>=0.15.0

# NLP
spacy>=3.7.2
nltk>=3.8.1
textblob>=0.17.1
vaderSentiment>=3.3.2
gensim>=4.3.2
sentence-transformers>=2.2.2
# en-core-web-trf@ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.2/en_core_web_trf-3.7.2-py3-none-any.whl

# Model optimization
optuna>=3.4.0
xgboost>=2.0.2
lightgbm>=4.1.0
mlflow>=2.8.1

# Error handling and resilience
circuitbreaker>=1.4.0
watchdog>=3.0.0

# Utilities
jinja2>=3.1.2
gunicorn>=21.2.0
pytz>=2023.3.post1
typing-extensions>=4.8.0
rich>=13.7.0
email-validator==2.1.0.post1
requests>=2.26.0
urllib3>=1.26.11
yarl>=1.17.0,<2.0
beautifulsoup4>=4.9.3
lxml>=5.4.0
networkx>=3.2.1
zlib-wrapper>=0.1.3

# Development and testing
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit>=3.5.0
locust>=2.8.3

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.14
mkdocstrings>=0.24.0

# Logging
structlog==23.2.0
python-json-logger==2.0.7 